Added BSSH Fastq copy handler to orcabus
Fixed a couple of ssm parameter structure issues with icav2 copy batch utility
