+++ config/config.ts
@@ -10,6 +10,7 @@ import { getSequenceRunManagerStackProps } from './stacks/sequenceRunManager';
 import { getFileManagerStackProps } from './stacks/fileManager';
 import { getBsRunsUploadManagerStackProps } from './stacks/bsRunsUploadManager';
 import { getICAv2CopyBatchUtilityStackProps } from './stacks/icav2CopyBatchUtility';
+import { getBsshIcav2FastqCopyManagerStackProps } from './stacks/bsshIcav2FastqCopyManager';
 
 interface EnvironmentConfig {
   name: string;
@@ -47,6 +48,8 @@ export const getEnvironmentConfig = (stage: AppStage): EnvironmentConfig | null 
             fileManagerStackProps: getFileManagerStackProps(stage),
             bsRunsUploadManagerStackProps: getBsRunsUploadManagerStackProps(stage),
             icav2CopyBatchUtilityStackProps: getICAv2CopyBatchUtilityStackProps(stage),
+            bsshIcav2FastqCopyManagerStackProps:
+              getBsshIcav2FastqCopyManagerStackProps(stage),
           },
         },
       };
@@ -69,6 +72,8 @@ export const getEnvironmentConfig = (stage: AppStage): EnvironmentConfig | null 
             fileManagerStackProps: getFileManagerStackProps(stage),
             bsRunsUploadManagerStackProps: getBsRunsUploadManagerStackProps(stage),
             icav2CopyBatchUtilityStackProps: getICAv2CopyBatchUtilityStackProps(stage),
+            bsshIcav2FastqCopyManagerStackProps:
+              getBsshIcav2FastqCopyManagerStackProps(stage),
           },
         },
       };
@@ -91,6 +96,8 @@ export const getEnvironmentConfig = (stage: AppStage): EnvironmentConfig | null 
             fileManagerStackProps: getFileManagerStackProps(stage),
             bsRunsUploadManagerStackProps: getBsRunsUploadManagerStackProps(stage),
             icav2CopyBatchUtilityStackProps: getICAv2CopyBatchUtilityStackProps(stage),
+            bsshIcav2FastqCopyManagerStackProps:
+              getBsshIcav2FastqCopyManagerStackProps(stage),
           },
         },
       };
