+++ config/constants.ts
@@ -1,4 +1,5 @@
 import { VpcLookupOptions } from 'aws-cdk-lib/aws-ec2';
+import path from 'path';
 
 export enum AppStage {
   BETA = 'beta',
@@ -106,6 +107,27 @@ export const icav2AccessTokenSecretName: Record<AppStage, string> = {
 Resources generated by the ICAv2 Copy Batch utility stack
  */
 export const icav2CopyBatchSSMRoot = '/orcabus/icav2_copy_batch_utility';
+export const icav2CopyBatchSSMBatchArn = path.join(icav2CopyBatchSSMRoot, 'batch_sfn_arn');
+export const icav2CopyBatchSSMBatchName = path.join(icav2CopyBatchSSMRoot, 'batch_sfn_name');
+export const icav2CopyBatchSSMSingleArn = path.join(icav2CopyBatchSSMRoot, 'single_sfn_arn');
+export const icav2CopyBatchSSMSingleName = path.join(icav2CopyBatchSSMRoot, 'single_sfn_name');
+export const icav2CopyBatchUtilityName = 'icav2_copy_batch_utility_sfn';
+export const icav2CopySingleUtilityName = 'icav2_copy_single_utility_sfn';
+
+/*
+Resources generated by the BSSH Fastq Copy Manager
+ */
+
+export const bsshFastqCopyManagerSSMRoot = '/orcabus/bssh_icav2_fastq_copy_manager';
+export const bsshFastqCopyManagerSSMName = path.join(
+  bsshFastqCopyManagerSSMRoot,
+  'state_machine_name'
+);
+export const bsshFastqCopyManagerSSMArn = path.join(
+  bsshFastqCopyManagerSSMRoot,
+  'state_machine_arn'
+);
+export const bsshFastqCopyManagerSfnName = 'bssh_fastq_copy_manager_sfn';
 
 // const statelessConfig = {
 //   multiSchemaConstructProps: {
