--- lib/workload/stateless/bssh_icav2_fastq_copy_manager/deploy/.npmignore
@@ -1,6 +1,0 @@
-*.ts
-!*.d.ts
-
-# CDK asset staging directory
-.cdk.staging
-cdk.out
