--- lib/workload/stateless/bssh_icav2_fastq_copy_manager/deploy/jest.config.js
@@ -1,8 +1,0 @@
-module.exports = {
-  testEnvironment: 'node',
-  roots: ['<rootDir>/test'],
-  testMatch: ['**/*.test.ts'],
-  transform: {
-    '^.+\\.tsx?$': 'ts-jest'
-  }
-};
