--- lib/workload/stateless/bssh_icav2_fastq_copy_manager/deploy/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "deploy",
-  "version": "0.1.0",
-  "bin": {
-    "deploy": "bin/deploy.js"
-  },
-  "scripts": {
-    "build": "tsc",
-    "watch": "tsc -w",
-    "test": "jest",
-    "cdk": "cdk"
-  },
-  "devDependencies": {
-    "@types/jest": "^29.5.11",
-    "@types/node": "20.11.5",
-    "jest": "^29.7.0",
-    "ts-jest": "^29.1.1",
-    "aws-cdk": "2.122.0",
-    "ts-node": "^10.9.2",
-    "typescript": "~5.3.3"
-  },
-  "dependencies": {
-    "aws-cdk-lib": "^2.122.0",
-    "constructs": "^10.0.0",
-    "source-map-support": "^0.5.21"
-  }
-}
