+++ lib/workload/stateless/stacks/bssh-icav2-fastq-copy-manager/deploy/test/deploy.test.ts
@@ -5,13 +5,12 @@
 // example test. To run these tests, uncomment this file along with the
 // example resource in lib/cttso_v2_launch_stack.ts
 test('SQS Queue Created', () => {
-//   const app = new cdk.App();
-//     // WHEN
-//   const stack = new Deploy.Bs_runs_upload_manager_stack(app, 'MyTestStack');
-//     // THEN
-//   const template = Template.fromStack(stack);
-
-//   template.hasResourceProperties('AWS::SQS::Queue', {
-//     VisibilityTimeout: 300
-//   });
+  //   const app = new cdk.App();
+  //     // WHEN
+  //   const stack = new Deploy.Bs_runs_upload_manager_stack(app, 'MyTestStack');
+  //     // THEN
+  //   const template = Template.fromStack(stack);
+  //   template.hasResourceProperties('AWS::SQS::Queue', {
+  //     VisibilityTimeout: 300
+  //   });
 });
