+++ lib/workload/stateless/stacks/bssh-icav2-fastq-copy-manager/images/step_functions_image.png
