+++ lib/workload/stateless/stacks/bssh-icav2-fastq-copy-manager/layers/pyproject.toml
