+++ lib/workload/stateless/statelessStackCollectionClass.ts
@@ -22,6 +22,10 @@ import {
   ICAv2CopyBatchUtilityStack,
   ICAv2CopyBatchUtilityStackProps,
 } from './stacks/icav2-copy-batch-utility/deploy/stack';
+import {
+  BsshIcav2FastqCopyManagerStack,
+  BsshIcav2FastqCopyManagerStackProps,
+} from './stacks/bssh-icav2-fastq-copy-manager/deploy/stack';
 
 export interface StatelessStackCollectionProps {
   postgresManagerStackProps: PostgresManagerStackProps;
@@ -30,6 +34,7 @@ export interface StatelessStackCollectionProps {
   fileManagerStackProps: FilemanagerProps;
   bsRunsUploadManagerStackProps: BsRunsUploadManagerStackProps;
   icav2CopyBatchUtilityStackProps: ICAv2CopyBatchUtilityStackProps;
+  bsshIcav2FastqCopyManagerStackProps: BsshIcav2FastqCopyManagerStackProps;
 }
 
 export class StatelessStackCollection {
@@ -40,6 +45,7 @@ export class StatelessStackCollection {
   readonly sequenceRunManagerStack: Stack;
   readonly bsRunsUploadManagerStack: Stack;
   readonly icav2CopyBatchUtilityStack: Stack;
+  readonly bsshIcav2FastqCopyManagerStack: Stack;
 
   constructor(
     scope: Construct,
@@ -81,6 +87,15 @@ export class StatelessStackCollection {
       {
         ...this.createTemplateProps(env, 'ICAv2CopyBatchUtilityStack'),
         ...statelessConfiguration.icav2CopyBatchUtilityStackProps,
+      }
+    );
+
+    this.bsshIcav2FastqCopyManagerStack = new BsshIcav2FastqCopyManagerStack(
+      scope,
+      'BsshIcav2FastqCopyManagerStack',
+      {
+        ...this.createTemplateProps(env, 'BsshIcav2FastqCopyManagerStack'),
+        ...statelessConfiguration.bsshIcav2FastqCopyManagerStackProps,
       }
     );
   }
